Added graphql structure instructions to readme.md
